requirements.txt
# AI 100 Final Project — Dependencies
# Authors: Jasraj "Jay" Raval & Jack Sweeney
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.5.0
